Update for next iteration of training
Presentation is moved out of repository. Only exercises remain in this repo.
The exercise have been updated according to the newer explanations in the presentation. In particular the naming of the parameters and variables have been made more consistent and the partial
derivatives are now the same as explained in the presentation.

Added another demo for using TensorFlow with the MNIST database. It shows a better actual usage of a neural network and can be used to combine everything into one final demo.

Bugs in exercises have been fixed. In particular the initialization of the weights was not done consistently with -0.5 and +0.5.

Cleaned up some unused parts.